Update intro-text
Over the last 25 years, the internet has shown us that the mishandling and abuse of personal data by third-parties is the biggest cyber security threat there is and that not sharing your data in the first place is the only way to truly secure it. Everyday, data breaches and hacks occur which threaten the safety, funds, reputation and identities of millions of internet users across the world. It’s time for change!

At Particl, we’re building a brand new decentralized and democratic economy supported by the Particl blockchain; a very secure and privacy-focused network where applications of all sorts can be built and operated without any middleman. Particl aims at offering the same range of services currently available on the web but with security, privacy, and most importantly, near-zero fees in mind.